Writing Data
Now let's write some data to Firebase.
You've done a great job of displaying messages, but to make this a real
chat room, we'll also need to allow the user to add messages to the
list. We've already taken care of binding the ng-keydown directive from
a text input field to a function called addMessage. You just have to
implement the addMessage function by appending new messages using the
$add method.

Just like Angular creates a 2-way data binding between a model and the
DOM, AngularFire creates a 2-way data binding between a model and
Firebase; so when you add a new message using $add, it also gets stored
into Firebase.